Account Management: refactoring operation class
