+++ FirebaseAuthUI/AccountManagement/FUIAccountSettingsOperationForgotPassword.h
@@ -17,4 +17,7 @@
 
 @interface FUIAccountSettingsOperationForgotPassword : FUIAccountSettingsOperation
 
++ (void)executeOperationWithDelegate:(id<FUIAccountSettingsOperationDelegate>)delegate
+                          showDialog:(BOOL)showDialog NS_UNAVAILABLE;
+
 @end
