+++ FirebaseAuthUI/AccountManagement/FUIAccountSettingsOperationSignOut.h
@@ -18,4 +18,6 @@
 
 @interface FUIAccountSettingsOperationSignOut : FUIAccountSettingsOperation
 
++ (void)executeOperationWithDelegate:(id<FUIAccountSettingsOperationDelegate>)delegate
+                          showDialog:(BOOL)showDialog NS_UNAVAILABLE;
 @end
